requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.0
pdfminer.six==20221105
reportlab==4.0.7
pytesseract==0.3.10
pillow==10.1.0
langchain==0.0.350
mistralai==0.1.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
